Successful creation of Sqlite database file, simple HTTP get operations.

+++ Code2Gether-Discord-Bot.Library/Models/ProjectManager.cs
@@ -47,9 +47,9 @@ namespace Code2Gether_Discord_Bot.Library.Models
             return _projectRepository.Update(project);
         }
 
-        private int GetNextId()
+        private long GetNextId()
         {
-            int i = 0;
+            long i = 0;
 
             try
             {
