+++ Code2Gether-Discord-Bot.Library/Models/Repositories/ProjectRepository/IProjectRepository.cs
@@ -5,10 +5,10 @@ namespace Code2Gether_Discord_Bot.Library.Models.Repositories.ProjectRepository
     public interface IProjectRepository
     {
         bool Create(Project newProject);
-        Project Read(int id);
+        Project Read(long id);
         Project Read(string projectName);
-        IDictionary<int, Project> ReadAll();
+        IDictionary<long, Project> ReadAll();
         bool Update(Project newProject);
-        bool Delete(int id);
+        bool Delete(long id);
     }
 }
