+++ Code2Gether-Discord-Bot.Library/Models/Repositories/ProjectRepository/ProjectDAL.cs
@@ -1,10 +1,7 @@
 ﻿using System;
 using System.Collections.Concurrent;
 using System.Collections.Generic;
-using System.Collections.Immutable;
 using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace Code2Gether_Discord_Bot.Library.Models.Repositories.ProjectRepository
 {
@@ -13,14 +10,14 @@ namespace Code2Gether_Discord_Bot.Library.Models.Repositories.ProjectRepository
     /// </summary>
     public class ProjectDAL : IProjectRepository
     {
-        private ConcurrentDictionary<int, Project> _projects = new ConcurrentDictionary<int, Project>();
+        private ConcurrentDictionary<long, Project> _projects = new ConcurrentDictionary<long, Project>();
 
         public bool Create(Project newProject)
         {
             return _projects.TryAdd(newProject.ID, newProject);
         }
 
-        public Project Read(int id)
+        public Project Read(long id)
         {
             if (_projects.TryGetValue(id, out Project project))
                 return project;
@@ -32,7 +29,7 @@ namespace Code2Gether_Discord_Bot.Library.Models.Repositories.ProjectRepository
             return ReadAll().Values.FirstOrDefault(p => p.Name == projectName);
         }
 
-        public IDictionary<int, Project> ReadAll()
+        public IDictionary<long, Project> ReadAll()
         {
             return _projects;
         }
@@ -42,7 +39,7 @@ namespace Code2Gether_Discord_Bot.Library.Models.Repositories.ProjectRepository
             return _projects.TryUpdate(newProject.ID, newProject, Read(newProject.ID));
         }
 
-        public bool Delete(int id)
+        public bool Delete(long id)
         {
             return _projects.TryRemove(id, out _);
         }
